add spring dependency; moved to package openweather
