--- src/main/java/de/ossi/model/forecast/Forecast.java
@@ -1,4 +1,0 @@
-package de.ossi.model.forecast;
-
-public record Forecast(City city) {
-}
