+++ src/main/java/de/ossi/openweather/OpenweatherApplication.java
@@ -1,0 +1,12 @@
+package de.ossi.openweather;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OpenweatherApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(OpenweatherApplication.class, args);
+    }
+}
