+++ src/main/java/de/ossi/openweather/model/currentweather/Main.java
@@ -1,8 +1,8 @@
-package de.ossi.model.currentweather;
+package de.ossi.openweather.model.currentweather;
 
 import com.google.gson.annotations.JsonAdapter;
 import com.google.gson.annotations.SerializedName;
-import de.ossi.json.KelvinToCelsiusConverter;
+import de.ossi.openweather.json.KelvinToCelsiusConverter;
 
 public record Main(@JsonAdapter(KelvinToCelsiusConverter.class) Double temp,
                    @JsonAdapter(KelvinToCelsiusConverter.class) @SerializedName("feels_like") Double feelsLike) {}
