+++ src/main/java/de/ossi/openweather/model/currentweather/Sys.java
@@ -1,4 +1,4 @@
-package de.ossi.model.currentweather;
+package de.ossi.openweather.model.currentweather;
 
 import java.time.LocalDateTime;
