+++ src/main/java/de/ossi/openweather/model/currentweather/Weather.java
@@ -1,3 +1,3 @@
-package de.ossi.model.currentweather;
+package de.ossi.openweather.model.currentweather;
 
 public record Weather(String main, String description) {}
