+++ src/main/java/de/ossi/openweather/model/forecast/City.java
@@ -1,4 +1,4 @@
-package de.ossi.model.forecast;
+package de.ossi.openweather.model.forecast;
 
 public record City(String name, String country) {
 }
