+++ src/main/java/de/ossi/openweather/model/forecast/Forecast.java
@@ -1,0 +1,4 @@
+package de.ossi.openweather.model.forecast;
+
+public record Forecast(City city) {
+}
