+++ src/test/java/de/ossi/openweather/Injectable.java
@@ -1,8 +1,8 @@
-package de.ossi;
+package de.ossi.openweather;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
-import de.ossi.injection.HttpClientModule;
+import de.ossi.openweather.injection.HttpClientModule;
 
 public interface Injectable {
     Injector injector = Guice.createInjector(new HttpClientModule());
